update setup service
