+++ src/i18n/locales/ja/features.ts
@@ -17,8 +17,8 @@ export const features = {
       desc: "予算、承認、振込プロセスを統合し、財務業務の効率を全面的に向上させます。",
     },
     analysis: {
-      title: "データ分析",
-      desc: "詳細な財務データ分析を提供し、経営層の賢明なビジネス上の意思決定を支援します。",
+      title: "顧問分析",
+      desc: "多分野の専門家による企業分析を提供し、経営陣の賢明なビジネス上の意思決定を支援します。",
     },
     tax: {
       title: "税務申告",
@@ -45,7 +45,7 @@ export const features = {
       desc: "外部の異種システムと柔軟に連携し、クロスプラットフォームの運用データをシームレスに統合します。",
     },
     esg: {
-      title: "温室効果ガス算定",
+      title: "カーボンインベントリ",
       desc: "活動データを正確に炭素排出量に換算し、企業のコンプライアンスリスクとコストを管理します。",
     },
     audit: {
