+++ src/i18n/locales/ko/features.ts
@@ -17,8 +17,8 @@ export const features = {
       desc: "예산, 승인 및 이체 프로세스를 통합하여 재무 업무 효율성을 종합적으로 향상시킵니다.",
     },
     analysis: {
-      title: "데이터 분석",
-      desc: "심층적인 재무 데이터 분석을 제공하여 경영진이 현명한 비즈니스 결정을 내릴 수 있도록 돕습니다.",
+      title: "자문 분석",
+      desc: "다양한 분야 전문가의 기업 분석을 제공하여 경영진이 현명한 비즈니스 결정을 내릴 수 있도록 지원합니다.",
     },
     tax: {
       title: "세금 신고",
@@ -45,7 +45,7 @@ export const features = {
       desc: "외부 이기종 시스템을 유연하게 연결하여 크로스 플랫폼 운영 데이터를 원활하게 통합합니다.",
     },
     esg: {
-      title: "온실가스 산정",
+      title: "탄소 인벤토리",
       desc: "활동 데이터를 탄소 배출량으로 정밀하게 환산하여 기업의 규정 준수 리스크와 비용을 관리합니다.",
     },
     audit: {
