+++ src/i18n/locales/zh_cn/admin_setup.ts
@@ -84,6 +84,15 @@ export const adminSetup = {
     deploying: "部署中",
     pending: "等待中",
     err_deploy: "部署合约失败。输出信息：",
+    use_existing_btn: "使用现有部署",
+    start_deployment_btn: "开始部署",
+    existing_found_title: "发现现有合约",
+    existing_found_desc: "如果您选择“使用现有部署”，下方应用的设置将不会生效。",
+    network_economics: "网络经济模型",
+    col_rate_desc: "配置绑定到此网络的默认抵押率（范围：100 ~ 10^-9）。",
+    col_rate_label: "ICP 抵押率 (1 ICP = X ISC)",
+    col_rate_placeholder: "例如 0.05",
+    col_rate_err: "抵押率必须介于 10^-9 和 100 之间。",
   },
   step5: {
     title: "步骤 5：初始化数据库",
