+++ src/i18n/locales/zh_tw/admin_setup.ts
@@ -85,6 +85,15 @@ export const adminSetup = {
     deploying: "部署中",
     pending: "等待中",
     err_deploy: "部署合約失敗。輸出訊息：",
+    use_existing_btn: "使用現有部署",
+    start_deployment_btn: "開始部署",
+    existing_found_title: "發現現有合約",
+    existing_found_desc: "如果您選擇「使用現有部署」，下方套用的設定將不會生效。",
+    network_economics: "網路經濟模型",
+    col_rate_desc: "設定綁定至此網路的預設抵押率（範圍：100 ~ 10^-9）。",
+    col_rate_label: "ICP 抵押率 (1 ICP = X ISC)",
+    col_rate_placeholder: "例如 0.05",
+    col_rate_err: "抵押率必須介於 10^-9 到 100 之間。",
   },
   step5: {
     title: "步驟 5：初始化資料庫",
