+++ src/i18n/locales/zh_tw/esg.ts
@@ -1,4 +1,5 @@
 export const esg = {
+  title: "碳盤查",
   detail_modal: {
     title: "碳盤查",
   },
